+++ src/App.css
@@ -1,0 +1,4 @@
+.App {
+  height: 100vh;
+  width: 100%;
+}
