+++ src/components/button/Button.css
@@ -1,0 +1,20 @@
+button {
+  border-radius: 100px;
+  height: 39px;
+  padding: 10px 24px 10px 24px;
+  cursor: pointer;
+  font-size: 16px;
+  font-weight: 500;
+}
+
+.button_1 {
+  color: #6d9F46;
+  border-color: #6d9F46;
+  background-color: transparent;
+}
+
+.button_2 {
+  color: #fff;
+  background-color: #6d9F46;
+  border-color: transparent;
+}
