+++ src/index.css
@@ -1,0 +1,8 @@
+html {
+  height: 100%;
+}
+
+body {
+  height: 100%;
+  margin: 0;
+}
