+++ src/pages/patient/SignIn.module.css
@@ -1,0 +1,55 @@
+.main_cont {
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  height: 100%;
+}
+
+.left_cont {
+  background-color: #c86f16;
+  width: 40%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.right_cont {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.form_cont {
+  margin-left: 120px;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  height: 80%;
+}
+
+.img {
+  width: 300px;
+  height: 350px;
+}
+
+.intro {
+  color: #fff;
+  font-size: 32px;
+  font-family: Roboto;
+  font-weight: 400;
+  width: 282px;
+  margin-top: 31.6%;
+}
+
+.heading {
+  font-size: 48px;
+  margin-top: 20%;
+  color: #f0851a;
+}
+
+.button_div {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin: 5px;
+}
